refactor(ui): move AsyncUpdateResultExt to dbflux_ui_base (#102)
Defines AsyncUpdateResultExt<T> in dbflux_ui_base::async_ext (pub trait),
re-exported from the crate root. Replaces the local definition in
dbflux_ui::ui::mod with a pub(crate) re-export so the ~9 call-sites
using crate::ui::AsyncUpdateResultExt remain unchanged.

dbflux_components remains anyhow-free (ADR-3).
