+++ crates/dbflux_ui_base/src/lib.rs
@@ -5,9 +5,11 @@
 //! full `dbflux_ui` crate.
 
 pub mod app_state_entity;
+pub mod async_ext;
 pub mod keymap;
 
 #[cfg(feature = "mcp")]
 pub use app_state_entity::McpRuntimeEventRaised;
 pub use app_state_entity::{AppStateChanged, AppStateEntity, AuthProfileCreated};
+pub use async_ext::AsyncUpdateResultExt;
 pub use keymap::{default_keymap, key_chord_from_gpui};
